@@ -34,14 +34,16 @@ namespace Helium.Controllers
         [HttpGet]
         [Produces("application/json")]
         [ProducesResponseType(typeof(string[]), 200)]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Globalization", "CA1303:Do not pass literals as localized parameters", Justification = "Method Name")]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "Method logs and handles all exceptions")]
         public async Task<IActionResult> GetGenresAsync()
         {
             // get list of genres as list of string
-            _logger.LogInformation("GetGenres");
+            _logger.LogInformation(nameof(GetGenresAsync));
 
             try
             {
-                return Ok(await _dal.GetGenresAsync());
+                return Ok(await _dal.GetGenresAsync().ConfigureAwait(false));
             }
 
             catch (CosmosException ce)
