@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -16,6 +17,6 @@ namespace Helium.DataAccessLayer
         Task<IEnumerable<Model.Movie>> GetMoviesAsync(int offset = 0, int limit = 0);
         Task<IEnumerable<Model.Movie>> GetMoviesByQueryAsync(string q, string genre = "", int year = 0, double rating = 0, bool toprated = false, string actorId = "", int offset = 0, int limit = 0);
         Task<List<string>> GetFeaturedMovieListAsync();
-        Task Reconnect(string cosmosUrl, string cosmosKey, string cosmosDatabase, string cosmosCollection, bool force = false);
+        Task Reconnect(Uri cosmosUrl, string cosmosKey, string cosmosDatabase, string cosmosCollection, bool force = false);
     }
 }
