@@ -11,10 +11,11 @@ namespace Helium.Model
     {
         public string Status { get; set; }
         public string ComponentType { get; set; }
-        public double ObservedValue { get; set; }
         public string ObservedUnit { get; set; }
+        public double ObservedValue { get; set; }
+        public double TargetValue { get; set; }
         public string Time { get; set; }
-        public List<string> AffectedEndpoints { get; set; }
+        public List<string> AffectedEndpoints { get; }
         public string Message { get; set; }
 
         /// <summary>
@@ -28,9 +29,15 @@ namespace Helium.Model
         /// <param name="hzCheck">HealthzCheck</param>
         public IetfCheck(HealthzCheck hzCheck)
         {
+            if (hzCheck == null)
+            {
+                throw new ArgumentNullException(nameof(hzCheck));
+            }
+
             Status = ToIetfStatus(hzCheck.Status);
             ComponentType = hzCheck.ComponentType;
             ObservedValue = Math.Round(hzCheck.Duration.TotalMilliseconds, 2);
+            TargetValue = Math.Round(hzCheck.TargetDuration.TotalMilliseconds, 0);
             ObservedUnit = "ms";
             Time = hzCheck.Time;
             Message = hzCheck.Message;
