@@ -1,5 +1,6 @@
 ﻿using Microsoft.Azure.KeyVault;
 using Microsoft.Extensions.DependencyInjection;
+using System;
 
 namespace KeyVault.Extensions
 {
@@ -12,7 +13,7 @@ namespace KeyVault.Extensions
         /// <param name="client">KeyVaultClient</param>
         /// <param name="uri">Key Vault URI</param>
         /// <returns>IServiceCollection</returns>
-        public static IServiceCollection AddKeyVaultConnection(this IServiceCollection services, KeyVaultClient client, string uri)
+        public static IServiceCollection AddKeyVaultConnection(this IServiceCollection services, KeyVaultClient client, Uri uri)
         {
             // add the KeyVaultConnection as a singleton
             services.AddSingleton<IKeyVaultConnection>(new KeyVaultConnection
