@@ -5,7 +5,7 @@ namespace Helium
 {
     public static class HomePageMiddlewareExtensions
     {
-        static readonly HashSet<string> validPaths = new HashSet<string> { "/", "/index.html", "/index.htm", "/default.html", "/default.htm" };
+        static readonly HashSet<string> validPaths = new HashSet<string> { "/", "/INDEX.HTML", "/INDEX.HTM", "/DEFAULT.HTML", "/DEFAULT.HTM" };
         static readonly byte[] responseBytes = System.Text.Encoding.UTF8.GetBytes("Under construction ...");
 
         /// <summary>
@@ -20,16 +20,16 @@ namespace Helium
             builder.Use(async (context, next) =>
             {
                 // matches / or index.htm[l] or default.htm[l]
-                if (validPaths.Contains(context.Request.Path.Value.ToLower()))
+                if (validPaths.Contains(context.Request.Path.Value.ToUpperInvariant()))
                 {
                     // return the content
                     context.Response.ContentType = "text/html";
-                    await context.Response.Body.WriteAsync(responseBytes, 0, responseBytes.Length);
+                    await context.Response.Body.WriteAsync(responseBytes, 0, responseBytes.Length).ConfigureAwait(false);
                 }
                 else
                 {
                     // not a match, so call next middleware handler
-                    await next();
+                    await next().ConfigureAwait(false);
                 }
             });
 
