@@ -18,7 +18,7 @@ namespace Helium
             builder.Use(async (context, next) =>
             {
                 // matches /version
-                if (context.Request.Path.Value.ToLower() == "/version")
+                if (context.Request.Path.Value.Equals("/version", System.StringComparison.OrdinalIgnoreCase))
                 {
                     if (_responseBytes == null)
                     {
@@ -27,12 +27,12 @@ namespace Helium
 
                     // return the content
                     context.Response.ContentType = "text/plain";
-                    await context.Response.Body.WriteAsync(_responseBytes, 0, _responseBytes.Length);
+                    await context.Response.Body.WriteAsync(_responseBytes, 0, _responseBytes.Length).ConfigureAwait(false);
                 }
                 else
                 {
                     // not a match, so call next middleware handler
-                    await next();
+                    await next().ConfigureAwait(false);
                 }
             });
 
